requirements-test.txt
# Test and Coverage Dependencies
coverage==7.3.2
coverage[toml]==7.3.2
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
